Temporarily withdraw `image/avif` from `MediaType`.
There is nothing wrong with it, and we will add it in version 33.5.0. I'm removing it merely so that we can release a [module-system-friendly](https://github.com/google/guava/issues/2970) version of Guava as a _patch_ release, rather than as a _minor_ release. (The `MediaType` change is the only addition we've made since 33.4.0.) By doing so, we reduce the pressure on users to upgrade immediately, since dependencies can't come to depend on any new APIs. That gives users a little more time to sort out any issues that might arise from our new `module-info`.

RELNOTES=n/a
PiperOrigin-RevId: 728267849