Override `Predicate.test` wherever we override `Predicate.apply`.
I went with `@since 21.0` because [that's when our `Predicate` began to extend the JDK's](https://github.com/google/guava/wiki/Release21#commonbase)—and thus it's the earliest release of Guava in which you can call, e.g., `someBloomFilter.test(someObject)`. I doubt that that will ever matter to anyone, though.

RELNOTES=n/a
PiperOrigin-RevId: 731937798