Test that `ForwardingExecutorServiceTest` does not forward the `default` method `close()`.
This has always been the case ([as it should be](https://github.com/google/guava/issues/6296#issuecomment-1385916649), given the goal of preserving historical behavior), and it's been documented since cl/502868842. But I figured it would be nice to have a test, especially given the R8 upgrade that we're preparing for in cl/736111584.

RELNOTES=n/a
PiperOrigin-RevId: 737693700