Use Javadoc `{@snippet ...}` tag.
This CL is mostly just the output of an automated search-and-replace operation:
- `<pre>{@code$` -> `{@snippet :`
- `[*] }</pre>` -> `* }`

But I did touch up `ComparisonChain`, which had a workaround for using "`@Override`" at the beginning of a line of the code snippet. In the process, I discovered that we hadn't carried over the "Java 8" instructions to the backport. Now that we use library desugaring for all our internal builds, we should include these instructions in the backport, too. I've done so. External users may or may not be able to use these features yet.

I also touched up `Iterables` and `FluentIterable`, since the `Class` overload of their `filter` methods was using a different style so that it could contain a line that starts with "`@SuppressWarnings`."

This CL should be safe now that [the external Guava build uses newer versions of tools like `javadoc`](https://github.com/google/guava/issues/6549), even as it continues to target (and test under) older JDKs.

The result of the CL is mostly that code snippets appear in a gray box with a copy button and slightly less whitespace.

RELNOTES=n/a
PiperOrigin-RevId: 737729648