Add `module-info` definition.
This changeset adds full support for modular Java builds in Guava and in libraries which depend on Guava.

The Guava JAR for JRE now structures as a Multi-Release JAR, with a module definition situated in `META-INF/versions/9/`. Guava remains compatible with JDK 8.

- Fixes https://github.com/google/guava/issues/2970
- Fixes https://github.com/google/guava/issues/7094

Relates-To: elide-dev/jpms#1
Signed-off-by: Sam Gammon <sam@elide.dev>
RELNOTES=Changed the Guava jar (and guava-testlib jar) to be a modular jar.
PiperOrigin-RevId: 738449919