Re-add `image/avif` in `MediaType`.
(And this time, include a `@since` tag :))

We'd originally added this in cl/723988216 but then withdrew it in cl/728267849 so that we could release a [module-system-friendly](https://github.com/google/guava/issues/2970) version of Guava as a _patch_ release, rather than as a _minor_ release.

RELNOTES=`net`: Added `image/avif` to `MediaType`.
PiperOrigin-RevId: 738775856