Fix a minor weirdness in the implementation of `ByteSource.contentEquals(ByteSource)` and move the part of its implementation that compares two `InputStream`s to a package-private method in `ByteStreams` while I'm at it.
The minor weirdness was this: when the same number of bytes was read into both buffers _but_ the number of bytes read was less than the length of the buffers, we would then use `Arrays.equals` to compare the _full_ contents of both buffers even though the reads only partially filled them.

This was not a correctness issue because it was always the case that if we got to that point, all previous reads into the buffers must have filled both buffers with the _same_ bytes, so the bytes in the section of the buffers that was not filled in the most recent read were guaranteed to be the same.

This was also not much of a performance issue since it could only occur once per `contentEquals` call, when the end of both `InputStream`s had been reached (the buffers are always completely filled otherwise).

That said, it's more obviously correct (and saves a tiny bit of work) to ensure that we only compare the bytes that were actually read in the most recent read operations.

RELNOTES=n/a
PiperOrigin-RevId: 738928305