Generate deterministic output for ambiguous short links.
The one that I'm aware of is `ForwardingBlockingDeque`. Our short link
keeps bouncing back and forth between the copy in `collect` and the copy
in `util.concurrent`:

https://github.com/google/guava/commits/gh-pages/javadocshortcuts/ForwardingBlockingDeque

I suspect that this commit will cause us to generate a link to the
non-deprecated one, the one in `util.concurrent`. But mainly I want to
give _some_ consistent answer.