Delete `ImmutableEntry` in favor of `SimpleImmutableEntry`.
Relevant to https://github.com/google/guava/issues/2861, but it's a step in the opposite direction (by not having a class for this at all, let alone a public one). (Alternatively: One reason not to provide a public `ImmutableEntry` type was that the current `ImmutableEntry` permits null values, which our other `Immutable*` types do not. From that perspective, this CL is a potential step _toward_ a public `ImmutableEntry` class, since it would allow us to reclaim the class name but with different behavior. But I wouldn't recommend holding your breath.)

Note that we already have an `ImmutableMapEntry` class ("an immutable `Map.Entry` for an `ImmutableMap`"), so it's nice to be rid of `ImmutableEntry` ("an immutable `Map.Entry` for any purpose).

RELNOTES=n/a
PiperOrigin-RevId: 778196681