Apply `private` to some classes that are nested directly inside a top-level class.
(We could apply it to some more deeply nested classes, too, but I don't want to go out of my way to do so: `private` would be redundant for any classes that are "effectively private.")

Plus, address a few Error Prone warnings in the backport-only `ObjectCountHashMap`. Notably, delete `ObjectCountHashMap.MapEntry.setCount`, which was introduced in cl/181210213 but never used (because we always use `ObjectCountHashMap.setValue`, I think).

RELNOTES=n/a
PiperOrigin-RevId: 792628431