In `Strings.lenientFormat`, don't assign into the varargs array.
When the method is called with an actual array argument, rather than as varargs, assigning into that array is at best surprising. It may also cause an `ArrayStoreException` if the component type of the array is not compatible with `String`.

RELNOTES=`Strings.lenientFormat` no longer assigns into its varargs array argument.
PiperOrigin-RevId: 800096793