Remove some unnecessary type arguments.
This is another CL that comes courtesy of IntelliJ, which even removed unnecessary type arguments _in Javadoc_.

I then went through the file and removed some unnecessary `Maps.` qualifiers from usages in this file. This isn't even a case of using static imports; this is just a case of not writing things like `Maps.keyFunction` when we're inside `Maps`. I did need to leave the qualifier in place for a nested class's `class KeySet extends Maps.KeySet<K, V>` for obvious reasons.

RELNOTES=n/a
PiperOrigin-RevId: 810148178