Standardize on the ID "`attach-javadocs`" for `maven-javadoc-plugin`.
Currently, our projects use a mix of that and "`attach-docs`." I think that the choice of ID is unlikely to make a difference in practice as long as we're consistent between a project and its parent (as was _not_ the case in jimfs prior to cl/814204365), but it's not hard to imagine messing things up someday if we copy and paste across projects.

"`attach-javadocs`" does seem to be at least [some kind of recommendation](https://maven.apache.org/plugins/maven-javadoc-plugin/faq.html#:~:text=attach%2Djavadocs).

Also, in Guava: In the `future` projects, remove some probably needless differences in configuration that we'd carried over from the main Guava projects back in the day:

```
<id>generate-javadoc-site-report</id>
<phase>site</phase>
```

I already removed that configuration from the main Guava projects back in https://github.com/google/guava/commit/6736cad8d092650d881ef0bada406ba843549ef1.

RELNOTES=n/a
PiperOrigin-RevId: 814338191