Further rework `ImmutableSet.copyOf` handling of `EnumSet` after https://github.com/google/guava/commit/83869af96382246da6ff99cb029ef59ba684f592.
I had thought (here _and_ during the review of https://github.com/google/guava/commit/6ab30067225eb1dd64e14aebdd0aec7f72739483): "Oh, maybe we could simplify this code because we don't _build under_ Java 8 anymore." But it looks like the problem we'd encountered was a problem with Java 8 _and higher_, not a problem that _went away after_ Java 8. (I found history in cl/98448037 and the threads linked from the review discussion there.) And I suspect that Java 8 is in the right. (Not that the code is "unsafe," just that we're doing things that look like they deserve the `unchecked` warnings that we're getting.)

It's possible that cl/98448037 went unnecessarily far in _also_ changing `ImmutableEnumSet` itself, after which we undid that part https://github.com/google/guava/commit/6ab30067225eb1dd64e14aebdd0aec7f72739483. Now I'm back for `ImmutableSet.copyOf` itself, with goals like:
- improve the suppression comment
- localize the suppression
- reduce diffs between mainline and backport
- maybe prefer `unchecked` warnings over `rawtypes` warnings, as the latter seem to have given us more trouble before (as in https://github.com/google/guava/commit/6ab30067225eb1dd64e14aebdd0aec7f72739483)

I'd originally intended to accomplish this by performing an unchecked cast to a set with arbitrary enum type (`Set<TimeUnit>`), operating on that, and then casting back to `ImmutableSet<E>` at the end. (`TimeUnit` wouldn't show up anywhere in the class file except in the local-variable table.) But after I got that working, I decided to ask Gemini, and it pointed out that we can do plenty with a wildcarded `EnumSet<?>`. I wonder if that approach didn't work at some point in the past, like when we actually did build under Java 8? Maybe I'm about to unwittingly break GWT/Eclipse? (If we encounter a problem specifically with `clone()`, I've had good results sticking with `EnumSet.copyOf` as an alternative.)

RELNOTES=n/a
PiperOrigin-RevId: 818031035