Update the `toString()` of some `CharMatcher` types.
Most `toString()` implementations mimic the code that could have produced the instance. Notably, `someCharMatcher.negate()` produces `someCharMatcher.toString() + ".negate()"`. But the `toString()` for `.and` and `.or` used a syntax that made it look as if these were static methods rather than instance ones.

Also update a couple of tests that depend on the exact `toString()` output.

RELNOTES=n/a
PiperOrigin-RevId: 829198847