Use version properties more, implicitly bumping `build-helper-maven-plugin` for `guava-tests`.
Compare https://github.com/google/guava/commit/3c0e7045d246aeafa2a12a21c72eb76aa1876768, in which I was too lazy to search for remaining versions. I've now grepped for `'<version>'` and looked for remaining non-property versions.

I'd been figuring that I should get around to this, but the impetus for doing it now is that I decided to ask Gemini an open-ended question about what could be improved in our `pom.xml` files, and this was its main recommendation. (It also suggested bumping the value in `requireMavenVersion` (but I might rather rip out that check entirely) and potentially removing the `plexus-io` workaround (but I'm not sure offhand if it's time yet).)

RELNOTES=n/a
PiperOrigin-RevId: 840494798