Inline `Serialization.readCount`.
The reason for that method's existence, according to cl/10687480, was so that we wouldn't have to mix calls to instance methods on `stream` with calls to helper methods on `Serialization`: We could instead use _exclusively_ methods from `Serialization`.

I'm not sure I would find that convincing on those merits, but I think we fall short even of that: We _still_ need to call an instance method on `stream`—namely, `defaultReadObject`. And at the time of cl/10687480, we were even mixing in a call to `stream.readInt()` in `ArrayListMultimap` (though we subsequently removed it).

(I noticed `readCount` while looking at `Serialization` as part of https://github.com/google/guava/issues/8099.)

RELNOTES=n/a
PiperOrigin-RevId: 840893040