Fix calling `remove` after `entry.setValue` on an iterator for `HashBiMap.inverse()`.
Also, add tests that apply not just to `HashBiMap` but to `Map` implementations in general.

Gemini noticed this bug during the review for cl/778043974. It observed that our code for the inverse-direction `BiMap` was missing this call that we have in the code for the "main" direction `BiMap`.

The bug appears to have existed since cl/35322418 in 2012.

RELNOTES=n/a
PiperOrigin-RevId: 845025562