Clear the fields of a `Node` inside `delete`.
This ensures that we don't retain memory to other `Node` instances and their contents, even in unlikely scenarios in which the user is holding a reference to one specific `Node` (through an `Entry`) but not to the others that it refers to.

This requires reading those fields in advance when replacing a node. But given that we already had complex ordering rules to follow in `putInverse`, I think that this is an improvement.

(This change was prompted by investigation in https://github.com/google/guava/commit/2097440502c5999e04d6cc5eac9ad15ffd971d86.)

RELNOTES=n/a
PiperOrigin-RevId: 846336997