Archive release Javadocs to make branch switches faster.
(I'm now seeing ~3s instead of ~30s to switch to `gh-pages` and ~2s instead of ~7s to switch back to `master`.)

I'd raised this possibility back in https://github.com/google/guava/issues/7597#issuecomment-2840049761.

This commit works in concert with a commit to the main branch to unarchive as part of building our site.